+++ app/(route)/(navbar)/search/page.tsx
@@ -1,0 +1,5 @@
+const SearchPage = () => {
+  return <div>추후 구현 될 검색 페이지</div>;
+};
+
+export default SearchPage;
